// src/app/api/model/cu-app-user-model.ts
/**
 * CrediFlow.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CUAppUserModel { 
    userId?: string | null;
    username: string;
    password?: string | null;
    fullName: string;
    phone?: string | null;
    email?: string | null;
    roleCode: string;
    storeId?: string | null;
    storeIds?: Array<string> | null;
    isActive?: boolean;
    mustChangePassword?: boolean;
}